Define universal data types.

+++ kernel.cpp
@@ -1,14 +1,15 @@
+#include "types.h"
 
 void printf(char* str) {
 
-  unsigned short* VideoMemory = (unsigned short*)0xb8000;
+  uint16_t* VideoMemory = (uint16_t*)0xb8000;
 
-  for(int i = 0; str[i]!= '\0'; ++i) {
+  for(uint16_t i = 0; str[i]!= '\0'; ++i) {
     VideoMemory[i] = (VideoMemory[i] & 0xff00) | str[i];
   }
 }
 
-extern "C" void kernelMain (void *multiboot_structure, unsigned int magic_number) {
+extern "C" void kernelMain (void *multiboot_structure, uint32_t magic_number) {
 
   printf("Hello Witcher!");
